+++ random_forest.py
@@ -1,95 +1,8 @@
 # Random forest algorithm
-import time
-from csv import reader
+
 from random import randrange
-from random import seed
-from math import sqrt
-from sklearn.ensemble import RandomForestClassifier
 
 
-
-####################################################################################################
-# class WineFileReader():
-#     def __init__(self, filename):
-#         raw_dataset = self.load_csv(filename)
-#         self.dataset = self.get_processed_dataset(raw_dataset)
-#         for row in self.dataset:
-#             class_id = row.pop()
-#
-#
-#     def load_csv(self, filename):
-#         dataset = []
-#         with open(filename, 'r') as file:
-#             csv_file = reader(file)
-#             for row in csv_file:
-#                 if row:
-#                     dataset.append(row)
-#         return dataset
-#
-#     def get_processed_dataset(self, raw_dataset):
-#         dataset_with_float_values = self.convert_values_to_float(raw_dataset)
-#         dataset = self.convert_class_name_to_int(dataset_with_float_values, len(dataset_with_float_values[0]) - 1)
-#         return dataset
-#
-#     def convert_values_to_float(self, raw_dataset):
-#         row_length = len(raw_dataset[0]) - 1
-#         for row in raw_dataset:
-#             for col in range(0, row_length):
-#                 row[col] = float(row[col].strip())
-#         return raw_dataset
-#
-#     def convert_class_name_to_int(self, dataset, column):
-#         class_values = [row[column] for row in dataset]
-#         unique = set(class_values)
-#         lookup = dict()
-#         for i, value in enumerate(unique):
-#             lookup[value] = i
-#         for row in dataset:
-#             row[column] = lookup[row[column]]
-#         return dataset
-
-
-##################################################################################################
-
-####################################################################################################
-class SonarFileReader():
-    def __init__(self, filename):
-        raw_dataset = self.load_csv(filename)
-        self.dataset = self.get_processed_dataset(raw_dataset)
-
-    def load_csv(self, filename):
-        dataset = []
-        with open(filename, 'r') as file:
-            csv_file = reader(file)
-            for row in csv_file:
-                if row:
-                    dataset.append(row)
-        return dataset
-
-    def get_processed_dataset(self, raw_dataset):
-        dataset_with_float_values = self.convert_values_to_float(raw_dataset)
-        dataset = self.convert_class_name_to_int(dataset_with_float_values, len(dataset_with_float_values[0]) - 1)
-        return dataset
-
-    def convert_values_to_float(self, raw_dataset):
-        row_length = len(raw_dataset[0]) - 1
-        for row in raw_dataset:
-            for col in range(0, row_length):
-                row[col] = float(row[col].strip())
-        return raw_dataset
-
-    def convert_class_name_to_int(self, dataset, column):
-        class_values = [row[column] for row in dataset]
-        unique = set(class_values)
-        lookup = dict()
-        for i, value in enumerate(unique):
-            lookup[value] = i
-        for row in dataset:
-            row[column] = lookup[row[column]]
-        return dataset
-
-
-##################################################################################################
 
 class DecisionTree():
     def __init__(self, dataset, max_depth, min_size, features_count):
@@ -142,17 +55,17 @@ class DecisionTreeNode():
         self.create_child_nodes()
 
     def init_node(self, dataset, features_count):  # M*N*sqrt(N) time, sqrt(N) + N + M^2*N*sqrt(N) space
-        class_values = list(set(row[-1] for row in dataset)) #N time, N space
+        class_values = list(set(row[-1] for row in dataset))  # N time, N space
         features = []
         range_end = len(dataset[0]) - 1
-        while len(features) < features_count:  #sqrt(N) time, sqrt(N) space
+        while len(features) < features_count:  # sqrt(N) time, sqrt(N) space
             random_id = randrange(range_end)
             if random_id not in features:
                 features.append(random_id)
         for index in features:
             for row in dataset:
-                tested_groups = get_potential_splitted_groups(index, row[index], dataset) #N time, N*M space
-                gini_index = get_gini_index(tested_groups, class_values) #N time, C space
+                tested_groups = get_potential_splitted_groups(index, row[index], dataset)  # N time, N*M space
+                gini_index = get_gini_index(tested_groups, class_values)  # N time, C space
                 if gini_index < self.split_score:
                     self.split_index = index
                     self.split_value = row[index]
@@ -249,9 +162,7 @@ class CustomRandomForestClassifier():
 
 # Helper methods#################################################################################
 
-
-# Calculate the Gini index for a split dataset
-def get_gini_index(groups, classes): #N time, C space
+def get_gini_index(groups, classes):  # N time, C space
     total_samples = float(sum([len(group) for group in groups]))
     gini_index = 0.0
     for group in groups:
@@ -265,9 +176,7 @@ def get_gini_index(groups, classes): #N time, C space
         gini_index += (1.0 - proportion_sum) * (size / total_samples)
     return gini_index
 
-
-# Split a dataset based on an attribute and an attribute value
-def get_potential_splitted_groups(index, value, dataset): #N time, N*M space
+def get_potential_splitted_groups(index, value, dataset):  # N time, N*M space
     left = []
     right = []
     for row in dataset:
@@ -276,85 +185,6 @@ def get_potential_splitted_groups(index, value, dataset): #N time, N*M space
         else:
             right.append(row)
     return left, right
-
-
-#######################################################################################################
-sonar_filename = 'data/sonar/sonar.all-data.csv'
-wine_filename = 'data/wine/wine.data.csv'
-max_depth = 10
-min_size = 1
-sample_size = 1.0
-seed(42)
-sonar_file_reader = SonarFileReader(sonar_filename)
-# wine_file_reader = WineFileReader(sonar_filename)
-sonar_dataset = sonar_file_reader.dataset
-# wine_dataset = wine_file_reader.dataset
-features_count_for_splitting = int(sqrt(len(sonar_dataset[0]) - 1))
-trees_count = 5
-
-
-# calculation accuracy ###############################################################################
-wins = 0
-sk_wins = 0
-count = len(sonar_dataset)
-speed_times = 0
-
-for i in range(count):
-    copy_dataset = list(sonar_dataset)
-    predicted_data = copy_dataset.pop(i)
-    start = time.time()
-    rf = CustomRandomForestClassifier(copy_dataset, max_depth, min_size, features_count_for_splitting, trees_count)
-    prediction = rf.predict(predicted_data)
-    end = time.time()
-    if prediction == predicted_data[-1]:
-        wins += 1
-
-
-    sk_dataset = [row[0:-1] for row in copy_dataset]
-    sk_classes = [row[-1] for row in copy_dataset]
-
-    sk_start = time.time()
-    clf = RandomForestClassifier(n_estimators=trees_count, max_depth=max_depth,
-                                 max_features=features_count_for_splitting)
-    clf.fit(sk_dataset, sk_classes)
-    sk_prediction = clf.predict([predicted_data[0:-1]])
-    sk_end = time.time()
-
-    if sk_prediction == predicted_data[-1]:
-       sk_wins += 1
-
-    speed_times += (end - start)/(sk_end - sk_start)
-
-
-print(speed_times/count)
-print('sk', 100 * sk_wins/count)
-print('my', 100 * wins/count)
-
-
-
-
-#
-# test_dataset = [[2.771244718, 1.784783929, 0],
-#                 [1.728571309, 1.169761413, 0],
-#                 [3.678319846, 2.81281357, 0],
-#                 [3.961043357, 2.61995032, 0],
-#                 [2.999208922, 2.209014212, 0],
-#                 [7.497545867, 3.162953546, 1],
-#                 [9.00220326, 3.339047188, 1],
-#                 [7.444542326, 0.476683375, 1],
-#                 [10.12493903, 3.234550982, 1],
-#                 [6.642287351, 3.319983761, 1]]
-# count = int(len(test_dataset) / 2)
-# for i in range(count):
-#     copy_dataset = list(test_dataset)
-#     predicted_data = copy_dataset.pop(i)
-#     rf = RandomForestClassifier(copy_dataset, max_depth, min_size, int(sqrt(len(test_dataset[0]) - 1)), 3)
-#     prediction = rf.predict(predicted_data)
-#     if prediction == predicted_data[-1]:
-#         wins += 1
-#
-# print(100 * wins/count)
-
